pyproject.toml
[project]
name = "cli-agant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.73.0",
    "gradio>=6.12.0",
    "groq>=1.1.2",
    "openai>=2.31.0",
    "python-dotenv>=1.2.2",
]
